Generate devices operations PSM
Generate devices operations PSM at the "PIM to PSM" command:
Moved the main part of this script (Translate procedure) to check some info before it is executed. Show an ERRORBOX if there's not a devices operations model related to the main model, or if there are no devices in the current model. If we are already on a Feature Operation Model there's no need to generate the PSM of the devices operations model.